// HrmsAngular/src/app/models/Attendance/shiftsetupModel.ts
export class ShiftSetupModel {

    id: number;
    shfitName: string;
    shiftStartHour: string;
    shiftStartMin: string;
    shiftEndtHour: string;
    shiftEndMin: string;
    minLogout: string;
    maxLogout: string;
    minIntime: string;
    maxIntime: string;
    out: number;
    nextDate: number;
    companyID: number;
    nextShiftID: number;
    dutyHours: number;
    dutyMinute: number;
}